zerolog adapter for std logger moved to own package
